requirements.txt
chromedriver-autoinstaller==0.5.1
numpy==1.23.5
pandas==1.5.3
psycopg2-binary==2.9.5
requests==2.28.2
beautifulsoup4==4.11.1
selenium==4.10.0
matplotlib==3.7.1
seaborn==0.12.2
python-telegram-bot==20.3
apache-airflow==2.6.3
apache-airflow-providers-postgres==4.0.2